+++ kolibri/plugins/style_guide/assets/src/views/StyleGuideIndex.vue
